give tags their own package
