+++ pkg/tags/tags_test.go
@@ -1,9 +1,27 @@
-package ocdav_test
+// Copyright 2018-2022 CERN
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// In applying this license, CERN does not waive the privileges and immunities
+// granted to it by virtue of its status as an Intergovernmental Organization
+// or submit itself to any jurisdiction.
+
+package tags_test
 
 import (
 	"testing"
 
-	"github.com/cs3org/reva/v2/internal/http/services/owncloud/ocdav"
+	"github.com/cs3org/reva/v2/pkg/tags"
 	"github.com/test-go/testify/require"
 )
 
@@ -37,11 +55,11 @@ func TestAddTags(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		tags := ocdav.FromString(tc.InitialTags)
+		ts := tags.FromString(tc.InitialTags)
 
-		added := tags.AddString(tc.TagsToAdd)
+		added := ts.AddString(tc.TagsToAdd)
 		require.Equal(t, tc.ExpectNoTagsAdded, !added)
-		require.Equal(t, tc.ExpectedTags, tags.AsString())
+		require.Equal(t, tc.ExpectedTags, ts.AsString())
 	}
 
 }
@@ -80,10 +98,10 @@ func TestRemoveTags(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		tags := ocdav.FromString(tc.InitialTags)
+		ts := tags.FromString(tc.InitialTags)
 
-		tags.RemoveString(tc.TagsToRemove)
-		require.Equal(t, tc.ExpectedTags, tags.AsString())
+		ts.RemoveString(tc.TagsToRemove)
+		require.Equal(t, tc.ExpectedTags, ts.AsString())
 	}
 
 }
